Clean up transitions, clamp initial rate to max.
Speed up the initial fade-in from 2s to 1s.
Don’t bother fading in the PIP when doing the flip.
Remove the timer that only hooked up the PIP halfway through, now that
Chrome handles backfaces properly.
Change the remaining timers to throw away the video on the backfaces
once the flip is complete.
Rework the case-insensitive fineLine search.
Clamp start bitrate to max bitrate.
Add the connect time to the infoDiv.